realign school viz component names with course viz components.
