+++ packages/frontend/app/components/school/session-type-visualize-vocabularies.hbs
@@ -31,6 +31,6 @@
     {{@model.title}}
   </h3>
   <div class="visualizations">
-    <School::VisualizerSessionTypeVocabularies @sessionType={{@model}} />
+    <School::VisualizerSessionTypeVocabulariesGraph @sessionType={{@model}} />
   </div>
 </section>
