+++ packages/frontend/app/components/school/session-type-visualize-vocabulary.hbs
@@ -37,7 +37,7 @@
     ({{@model.vocabulary.title}})
   </h3>
   <div class="visualizations">
-    <School::VisualizerSessionTypeVocabulary
+    <School::VisualizerSessionTypeVocabularyGraph
       @sessionType={{@model.sessionType}}
       @vocabulary={{@model.vocabulary}}
     />
