+++ packages/frontend/app/components/school/visualizer-session-type-vocabulary-graph.hbs
@@ -1,6 +1,6 @@
 <span
-  class="school-visualizer-session-type-vocabularies not-icon"
-  data-test-school-visualizer-session-type-vocabularies
+  class="{{unless @isIcon 'not-icon'}} school-visualizer-session-type-vocabulary-graph"
+  data-test-school-visualizer-session-type-vocabulary-graph
   ...attributes
 >
   {{#if this.isLoaded}}
@@ -9,7 +9,6 @@
         @name="donut"
         @isIcon={{@isIcon}}
         @data={{this.data}}
-        @onClick={{this.donutClick}}
         @hover={{perform this.donutHover}}
         @leave={{perform this.donutHover}}
         as |chart|
