+++ packages/frontend/app/styles/components.scss
@@ -153,8 +153,8 @@
 
 @forward "components/school/session-type-visualize-vocabularies";
 @forward "components/school/session-type-visualize-vocabulary";
-@forward "components/school/visualizer-session-type-vocabularies";
-@forward "components/school/visualizer-session-type-vocabulary";
+@forward "components/school/visualizer-session-type-vocabularies-graph";
+@forward "components/school/visualizer-session-type-vocabulary-graph";
 @forward "components/school/emails";
 @forward "components/school/emails-editor";
