+++ packages/frontend/app/styles/components/school/visualizer-session-type-vocabularies-graph.scss
@@ -1,4 +1,4 @@
-.school-visualizer-session-type-vocabularies {
+.school-visualizer-session-type-vocabularies-graph {
   display: inline-block;
   height: 1rem;
   width: 1rem;
