+++ packages/frontend/app/styles/components/school/visualizer-session-type-vocabulary-graph.scss
@@ -1,4 +1,4 @@
-.school-visualizer-session-type-vocabulary {
+.school-visualizer-session-type-vocabulary-graph {
   display: inline-block;
   height: 1rem;
   width: 1rem;
